Allow raw pixel scrolling in web contents in Aura.
A future patch will allow raw pixel scrolling in views.

ui::LocatedEvent stores a floating point location. Currently
integral data is returned for all previously existing
accessors. In the future, all accessors should return the
floating point values.

BUG=329354
TEST=GestureRecognizerTest.GestureEventScroll

Review URL: https://codereview.chromium.org/143603002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@247358 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: e53cc7742f11156650dd63362383eb6cec27e0cd

+++ ui/events/gestures/velocity_calculator.cc
@@ -30,7 +30,7 @@ float VelocityCalculator::YVelocity() {
   return y_velocity_;
 }
 
-void VelocityCalculator::PointSeen(int x, int y, int64 time) {
+void VelocityCalculator::PointSeen(float x, float y, int64 time) {
   buffer_[index_].x = x;
   buffer_[index_].y = y;
   buffer_[index_].time = time;
